requirements.txt
django
djangorestframework
django-cors-headers
psycopg[binary]
drf-spectacular
djangorestframework-simplejwt
google-auth
django-apscheduler